requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
Django==3.1.2
django-embed-video==1.3.3
idna==2.10
numpy==1.19.2
opencv-python==4.4.0.44
Pillow==8.0.1
psycopg2==2.8.6
pytz==2020.1
requests==2.24.0
sqlparse==0.4.1
urllib3==1.25.11
